docs: use react template for toptask-nav
