+++ components/toptask-nav/css/stories/default.stories.mdx
@@ -5,25 +5,15 @@ Copyright (c) 2021 Gemeente Utrecht
 -->
 
 import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
-import { defaultArgs, exampleLinks, ToptaskNav } from "../template";
+import { argTypes, defaultArgs, exampleArgs, ToptaskNav } from "../story-template";
 import "../index.scss";
 
 <Meta
   id="css-toptask-nav"
   title="CSS Component/Toptask Navigation"
-  argTypes={{
-    links: {
-      description: "Links",
-      type: {
-        name: "array",
-        required: true,
-      },
-    },
-    grid: {
-      type: "boolean",
-      description: "Grid",
-    },
-  }}
+  component={ToptaskNav}
+  argTypes={argTypes}
+  args={defaultArgs}
   parameters={{
     status: {
       type: "WORK IN PROGRESS",
@@ -38,10 +28,7 @@ Styling via the `.utrecht-toptask-nav` class name:
 <Canvas>
   <Story
     name="Default"
-    args={{
-      ...defaultArgs,
-      links: [...exampleLinks],
-    }}
+    args={exampleArgs}
     parameters={{
       viewport: {
         viewports: {
